students/k3340/Yakshin_Artemiy/Lr_1/lab_1/requirements.txt
fastapi[all]
sqlmodel
psycopg2-binary
alembic
python-dotenv
bcrypt
PyJWT
typing_extensions
